# internal/api/xmproject.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.15.0 DO NOT EDIT.
package api

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for CompanyType.
const (
	Cooperative        CompanyType = "cooperative"
	Corporations       CompanyType = "corporations"
	NonProfit          CompanyType = "non_profit"
	SoleProprietorship CompanyType = "sole_proprietorship"
)

// Company defines model for Company.
type Company struct {
	Description    *string     `json:"description,omitempty"`
	EmployeesCount int         `json:"employees_count"`
	Id             string      `json:"id"`
	Name           string      `json:"name"`
	Registered     bool        `json:"registered"`
	Type           CompanyType `json:"type"`
}

// CompanyType defines model for CompanyType.
type CompanyType string

// CreateCompany defines model for CreateCompany.
type CreateCompany struct {
	Description    *string     `json:"description,omitempty"`
	EmployeesCount int         `json:"employees_count"`
	Name           string      `json:"name"`
	Registered     bool        `json:"registered"`
	Type           CompanyType `json:"type"`
}

// Error defines model for Error.
type Error struct {
	Error string `json:"error"`
}

// IdResponse defines model for IdResponse.
type IdResponse struct {
	Id string `json:"id"`
}

// UpdateCompany defines model for UpdateCompany.
type UpdateCompany struct {
	Description    *string      `json:"description,omitempty"`
	EmployeesCount *int         `json:"employees_count,omitempty"`
	Name           *string      `json:"name,omitempty"`
	Registered     *bool        `json:"registered,omitempty"`
	Type           *CompanyType `json:"type,omitempty"`
}

// CreateCompanyJSONRequestBody defines body for CreateCompany for application/json ContentType.
type CreateCompanyJSONRequestBody = CreateCompany

// UpdateCompanyJSONRequestBody defines body for UpdateCompany for application/json ContentType.
type UpdateCompanyJSONRequestBody = UpdateCompany